requirements.txt
docker
GitPython
boto3
botocore
python-dotenv